requirements.txt
# Core Framework
Flask==3.0.0
Werkzeug==3.0.1

# Database
SQLAlchemy==2.0.35
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5

# HTTP requests (for AI features)
requests==2.32.3

# Environment & Configuration
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# Development
Flask-DebugToolbar==0.14.1

# AI/ML
openai==1.58.1

# Production Server
gunicorn==21.2.0

# HTML Parsing (for tests)
beautifulsoup4==4.12.3
